Move stateless methods out of composable
